+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed a PHP error that could occur when saving a feed using an Asset element type with no volume selected.
 - Fixed a PHP error that could occur when importing a feed  that has `setEmptyValues` set to off on the feed. ([#1269](https://github.com/craftcms/feed-me/issues/1269))
+- Fixed several bugs related to empty and non-existent feed values and the “Set Empty Values” feed setting. ([#1271](https://github.com/craftcms/feed-me/pull/1271))
 
 ## 5.1.1.1 - 2023-02-24